infrastructure/files/requirements.txt
google-cloud-storage
google-api-core
google-cloud-discoveryengine
google-cloud-logging
